remove ruleCondition
