+++ weevent-governance/src/main/java/com/webank/weevent/governance/entity/RuleEngineEntity.java
@@ -64,7 +64,6 @@ public class RuleEngineEntity extends RuleEngineBase {
     @Transient
     private String tableName;
 
-
-    @Transient
-    private List<RuleEngineConditionEntity> ruleConditionList = new ArrayList<>();
 }
+
+
