+++ weevent-governance/src/main/java/com/webank/weevent/governance/utils/JsonUtil.java
@@ -1,14 +1,10 @@
 package com.webank.weevent.governance.utils;
 
 import java.io.IOException;
-import java.util.List;
 import java.util.Map;
-
-import com.webank.weevent.governance.entity.RuleEngineConditionEntity;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.type.MapLikeType;
@@ -58,19 +54,6 @@ public class JsonUtil {
         } catch (Exception e) {
             log.error("conversion of Json failed", e);
             throw new IOException("conversion of Json failed", e);
-        }
-    }
-
-    public static <T> List<T> json2List(String jsonString) throws IOException {
-        try {
-            if (jsonString == null) {
-                return null;
-            }
-            return objectMapper.readValue(jsonString, new TypeReference<List<RuleEngineConditionEntity>>() {
-            });
-        } catch (JsonProcessingException e) {
-            log.error("convert jsonString to List failed ", e);
-            throw new IOException("conversion of List failed", e);
         }
     }
