+++ weevent-governance/src/main/resources/script/governance_h2.sql
@@ -58,30 +58,23 @@ create TABLE t_account (
 	broker_id INTEGER,
 	cep_id VARCHAR (255),
 	condition_field VARCHAR (4096),
+	condition_field_json VARCHAR (4096),
 	condition_type INTEGER,
 	delete_at BIGINT NOT NULL,
 	error_destination VARCHAR (255),
 	from_destination VARCHAR (255),
 	function_array VARCHAR (1024),
 	group_id VARCHAR (255),
-	payload VARCHAR (255),
+	payload VARCHAR (4096),
 	payload_type INTEGER,
 	rule_database_id INTEGER,
 	rule_name VARCHAR (128),
-	select_field VARCHAR (255),
+	select_field VARCHAR (4096),
 	STATUS INTEGER,
 	system_tag BOOLEAN,
 	to_destination VARCHAR (255),
 	user_id INTEGER,
 	brokerUrl VARCHAR (255),
-	PRIMARY KEY (id)
-);
- create TABLE t_rule_engine_condition (
-	id INTEGER generated BY DEFAULT AS identity,
-	create_date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-	last_update TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-	rule_id INTEGER,
-	sql_condition_json VARCHAR (255),
 	PRIMARY KEY (id)
 );
  create TABLE t_topic (
